Install SSO on Pharo 8
Even though its old, this page was prominent in some of of the searches I was doing on... Pharo Zinc SSO,
and since I'm out of practice with Pharo it mislead me for a few hours.  Hope this helps the next person.

P.S. I'm finding it hard to find recent tutorial on SSO use.  Any you can refer me to?
cheers -ben